Fix import paths on moved unit tests
